requirements.txt
numpy==1.19.5
tensorflow==1.13.2
Keras==2.2.0
h5py==2.10.0
matplotlib==3.1.0
scikit-learn==0.24.2
opencc-python-reimplemented==0.1.6
ckiptagger==0.2.1
gdown==4.2.0